+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Adjusted `Endpoint::body` to return an `Option` on success
 - Require `std::error::Error` instead of `std::fmt::Display` for
   `Endpoint::ApiError` type
 - Bumped minimum supported Rust version to `1.46`